+++ test/PHPUnit/model/DocumentActionEntiteTest.php
@@ -49,5 +49,37 @@ class DocumentActionEntiteTest extends PastellTestCase {
 		);
 	}
 
+	/**
+	 * @throws Exception
+	 */
+	public function testGetDocumentOlderThanDay(){
+		$id_d = $this->createDocument();
+		$this->addAction($id_d,"action-test");
+		$this->addAction($id_d,"modification");
+		$documents_list = $this->documentActionEntite->getDocumentOlderThanDay(
+			1,
+			"document-type-test",
+			"action-test",
+			0
+		);
+		$this->assertEquals([],$documents_list);
+	}
+
+	/**
+	 * @throws Exception
+	 */
+	public function testGetDocumentInStateOlderThanDay(){
+		$id_d = $this->createDocument();
+		$this->addAction($id_d,"action-test");
+		$this->addAction($id_d,"modification");
+		$documents_list = $this->documentActionEntite->getDocumentInStateOlderThanDay(
+			1,
+			"document-type-test",
+			"action-test",
+			0
+		);
+		$this->assertEquals("modification",$documents_list[0]['last_action']);
+	}
+
 }